feat(models_web): add permission checking to create/update/delete

+++ models_WEB/teacher.py
@@ -5,7 +5,8 @@ from datetime import datetime
 from flask_restful import Resource, request
 
 from database.mssql import conn, cursor
-from frontend.helper_web.MESSAGE import ACTION_MUST_BE_CRUD, CREATE_GENERAL_MSG, MISSING_ARGS_MSG
+from frontend.helper_web.have_permission import have_permission
+from frontend.helper_web.MESSAGE import ACTION_MUST_BE_CRUD, CREATE_GENERAL_MSG, INVALID_ROLE, MISSING_ARGS_MSG
 from frontend.helper_web.validate_args import validate_args
 
 
@@ -57,16 +58,20 @@ class TeacherAPI(Resource):
 
     def CREATE(self):
         validate_success, message_body, missing_args = validate_args(
-            request.get_json(silent=True), tuple(["teacher_id", "teacher_name", "date_of_birth", "email"])
+            request.get_json(silent=True), tuple(["teacher_id", "teacher_name", "date_of_birth", "email", "session_key"])
         )
         if not validate_success:
             return {"message": MISSING_ARGS_MSG(missing_args), "data": {}}, 400
-        teacher_id, teacher_name, date_of_birth_str, email = (
+        teacher_id, teacher_name, date_of_birth_str, email, session_key = (
             message_body["teacher_id"],
             message_body["teacher_name"],
             message_body["date_of_birth"],
             message_body["email"],
+            message_body["session_key"],
         )
+        if not have_permission(session_key, admin_only=True):
+            return {"message": INVALID_ROLE, "data": {}}, 403
+
         if (teacher_id == "") or (teacher_name == "") or (date_of_birth_str == "") or (email == ""):
             return {"message": "all fields cannot be empty", "data": {}}, 400
 
@@ -93,16 +98,20 @@ class TeacherAPI(Resource):
 
     def UPDATE(self):
         validate_success, message_body, missing_args = validate_args(
-            request.get_json(silent=True), tuple(["teacher_id", "teacher_name", "date_of_birth", "email"])
+            request.get_json(silent=True), tuple(["teacher_id", "teacher_name", "date_of_birth", "email", "session_key"])
         )
         if not validate_success:
             return {"message": MISSING_ARGS_MSG(missing_args), "data": {}}, 400
-        teacher_id, teacher_name, date_of_birth, email = (
+        teacher_id, teacher_name, date_of_birth, email, session_key = (
             message_body["teacher_id"],
             message_body["teacher_name"],
             message_body["date_of_birth"],
             message_body["email"],
+            message_body["session_key"],
         )
+        if not have_permission(session_key, admin_only=True):
+            return {"message": INVALID_ROLE, "data": {}}, 403
+
         if teacher_id == "":
             return {"message": "teacher_id cannot be empty", "data": {}}, 400
 
@@ -134,10 +143,14 @@ class TeacherAPI(Resource):
         )
 
     def DELETE(self):
-        validate_success, message_body, missing_args = validate_args(request.get_json(silent=True), tuple(["teacher_id"]))
+        validate_success, message_body, missing_args = validate_args(
+            request.get_json(silent=True), tuple(["teacher_id", "session_key"])
+        )
         if not validate_success:
             return {"message": MISSING_ARGS_MSG(missing_args), "data": {}}, 400
-        teacher_id = message_body["teacher_id"]
+        teacher_id, session_key = message_body["teacher_id"], message_body["session_key"]
+        if not have_permission(session_key, admin_only=True):
+            return {"message": INVALID_ROLE, "data": {}}, 403
 
         cursor.execute(f"DELETE FROM Teachers WHERE TeacherID = '{teacher_id}'")
         conn.commit()
